requirements.txt
openai
google.generativeai
langchain
langchain_community
pandas
matplotlib
seaborn
tiktoken
chromadb
pypdf
docx
python-docx
streamlit
